+++ src/main/java/blackjack/domain/BlackJackGame.java
@@ -3,6 +3,8 @@ package blackjack.domain;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
 import blackjack.domain.card.generator.DeckGenerator;
+import blackjack.domain.result.GameResult;
+import blackjack.domain.result.Score;
 import blackjack.domain.user.Dealer;
 import blackjack.domain.user.Users;
 import blackjack.dto.CardAndScoreResult;
