+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 public enum GameResult {
     WIN,
