+++ src/main/java/blackjack/dto/DealerFinalResult.java
@@ -1,6 +1,6 @@
 package blackjack.dto;
 
-import blackjack.domain.GameResult;
+import blackjack.domain.result.GameResult;
 
 import java.util.Map;
