+++ src/main/java/blackjack/view/ViewRenderer.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.GameResult;
+import blackjack.domain.result.GameResult;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
