+++ src/test/java/blackjack/domain/card/CardGroupTest.java
@@ -1,6 +1,6 @@
 package blackjack.domain.card;
 
-import blackjack.domain.Score;
+import blackjack.domain.result.Score;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
