+++ src/test/java/blackjack/domain/card/generator/BlackJackGameTest.java
@@ -1,7 +1,7 @@
 package blackjack.domain.card.generator;
 
 import blackjack.domain.BlackJackGame;
-import blackjack.domain.GameResult;
+import blackjack.domain.result.GameResult;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
