+++ src/test/java/blackjack/domain/user/DealerTest.java
@@ -1,7 +1,7 @@
 package blackjack.domain.user;
 
-import blackjack.domain.GameResult;
-import blackjack.domain.Score;
+import blackjack.domain.result.GameResult;
+import blackjack.domain.result.Score;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardGroup;
 import blackjack.domain.card.CardNumber;
